+++ src/Http/LogoutModule.php
@@ -56,7 +56,7 @@ class LogoutModule implements ServiceModuleInterface
                     $this->session->delete('_two_factor_enroll_redirect_to');
                     $this->session->delete('_two_factor_verified');
                     $this->session->delete('_form_auth_user');
-                    $this->session->delete('_form_auth_entitlement_list');
+                    $this->session->delete('_form_auth_permission_list');
                     $this->session->delete('_form_auth_time');
 
                     // a logout URL is defined, this is used by SAML/Mellon